refactor(youtube-downloader): use local yt-dlp binary instead of Docker
- Replace Docker-based ytdlp helper with direct yt-dlp calls (ytdlp.ts)
- Update download, extract-playlist, and check-zzz-playlists scripts
- Remove ytmp3.dockerfile and the Docker image build/run logic
- Document yt-dlp/ffmpeg requirement in agents.md
