+++ agents.md
@@ -10,9 +10,10 @@ Never translate. We already have a script that do that yarn i18n.
 - Standard commands (in `package.json`): `yarn test` (Vitest), `yarn typecheck` (tsc), `yarn build` (production Gatsby build), `yarn serve` (serve built output). There is no lint script/ESLint config; `yarn typecheck` is the closest static-analysis gate.
 - Known pre-existing failures (not environment issues): `yarn test` has 1 failing `addHash` case, and `yarn typecheck` reports errors in `src/components/ui/resizable.tsx`, `src/components/ui/skeleton.tsx`, and `src/hooks/use-toast.ts`. `yarn build` still succeeds.
 
-### Music helper scripts (Docker / ffmpeg)
+### Music helper scripts (yt-dlp / ffmpeg)
 
-- `./scripts/get-music-durations.sh` needs `ffprobe` (ffmpeg), which is installed system-wide — it works directly, no Docker needed.
-- `yarn mp3` (`scripts/youtube-downloader/download.ts`), `yarn playlist`, and `yarn check-zzz-playlists` require **Docker**: they build/run the `zzz-ytmp3` image (`scripts/youtube-downloader/ytmp3.dockerfile`, a `yt-dlp` + ffmpeg container) and call the `docker` CLI directly.
-- The Docker daemon does not auto-start (no systemd here). Start it once with `sudo dockerd` (e.g. in a background tmux session). The `ubuntu` user is in the `docker` group, so `docker`/`yarn mp3` run without `sudo` in a fresh login shell; in an already-open shell, prefix with `sg docker -c '...'`.
-- `yarn mp3` reads URLs from `scripts/youtube-downloader/files-to-download.txt` and writes MP3s to `scripts/youtube-downloader/files/`. From a cloud/datacenter IP, YouTube often returns "Sign in to confirm you're not a bot"; downloads then need cookies (`--cookies`/`--cookies-from-browser`). This is a YouTube-side limitation, not an environment problem.
+- These scripts call the local `yt-dlp` binary directly (no Docker). `ffmpeg`/`ffprobe` is installed system-wide.
+- `./scripts/get-music-durations.sh` uses `ffprobe`.
+- `yarn mp3` (`scripts/youtube-downloader/download.ts`), `yarn playlist`, and `yarn check-zzz-playlists` use `yt-dlp` via the shared helper `scripts/youtube-downloader/ytdlp.ts`. `yt-dlp` must be on `PATH` (the update script installs it to `/usr/local/bin`).
+- `yarn mp3` reads URLs from `scripts/youtube-downloader/files-to-download.txt` and writes MP3s to `scripts/youtube-downloader/files/`.
+- From a cloud/datacenter IP, YouTube often returns "Sign in to confirm you're not a bot"; downloads then need cookies (`--cookies`/`--cookies-from-browser`), and recent `yt-dlp` may also warn about needing a JS runtime (e.g. deno). These are YouTube/yt-dlp limitations, not environment problems — non-YouTube direct media URLs download fine.
