+++ scripts/youtube-downloader/download.ts
@@ -1,28 +1,20 @@
 import { promises as fs } from "fs";
 import path from "path";
-import { spawn } from "child_process";
 import { fileURLToPath } from "url";
+import { requireCmd, runCommand, ytdlpBin } from "./ytdlp";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
 const scriptDir = __dirname;
-const repoRoot = path.resolve(scriptDir, "..", "..");
 
 const filesList = path.join(scriptDir, "files-to-download.txt");
 const outDir = path.join(scriptDir, "files");
-const dockerfilePath = path.join(
-  repoRoot,
-  "scripts/youtube-downloader",
-  "ytmp3.dockerfile"
-);
-const imageName = "zzz-ytmp3";
 
 function printUsage() {
-  // Keep usage text close to the original Bash script
   console.log(
     [
-      "YouTube to MP3 downloader (Docker-based)",
+      "YouTube to MP3 downloader (yt-dlp)",
       "",
       "Usage:",
       `  ${path.basename(process.argv[1] || "download.ts")}`,
@@ -30,7 +22,7 @@ function printUsage() {
       "Behavior:",
       `  - Reads URLs from: ${filesList}`,
       `  - Writes MP3 files to: ${outDir}`,
-      `  - Uses Docker image: ${imageName}`,
+      `  - Uses the local '${ytdlpBin}' binary`,
       "",
       "Notes:",
       "  - One URL per line.",
@@ -39,81 +31,13 @@ function printUsage() {
   );
 }
 
-async function runCommand(
-  cmd: string,
-  args: string[],
-  options: { silent?: boolean } = {}
-): Promise<number> {
-  return new Promise((resolve, reject) => {
-    const child = spawn(cmd, args, {
-      stdio: options.silent ? "ignore" : "inherit",
-    });
-
-    child.on("error", (err) => {
-      reject(err);
-    });
-
-    child.on("close", (code) => {
-      resolve(code ?? 1);
-    });
-  });
-}
-
-async function requireCmd(cmd: string) {
-  try {
-    const code = await runCommand(cmd, ["--version"], { silent: true });
-    if (code !== 0) {
-      throw new Error();
-    }
-  } catch {
-    console.error(`Error: '${cmd}' is not installed or not in PATH.`);
-    process.exit(1);
-  }
-}
-
-async function ensureDockerImage() {
-  // Check if image exists
-  const inspectCode = await runCommand(
-    "docker",
-    ["image", "inspect", imageName],
-    { silent: true }
-  );
-
-  if (inspectCode === 0) {
-    return;
-  }
-
-  // Build image if missing
-  try {
-    await fs.access(dockerfilePath);
-  } catch {
-    console.error(`Error: Dockerfile not found at ${dockerfilePath}`);
-    process.exit(1);
-  }
-
-  console.log(`Building Docker image '${imageName}'...`);
-  const buildCode = await runCommand("docker", [
-    "build",
-    "-t",
-    imageName,
-    "-f",
-    dockerfilePath,
-    repoRoot,
-  ]);
-
-  if (buildCode !== 0) {
-    console.error("Error: Failed to build Docker image.");
-    process.exit(1);
-  }
-}
-
 async function main() {
   if (process.argv[2] === "-h" || process.argv[2] === "--help") {
     printUsage();
     return;
   }
 
-  await requireCmd("docker");
+  await requireCmd(ytdlpBin);
 
   await fs.mkdir(outDir, { recursive: true });
 
@@ -139,8 +63,6 @@ async function main() {
     process.exit(1);
   }
 
-  await ensureDockerImage();
-
   console.log(`Output directory: ${outDir}`);
   console.log(`Total URLs: ${numUrls}`);
   console.log();
@@ -152,12 +74,7 @@ async function main() {
   for (const url of urls) {
     console.log(`Downloading: ${url}`);
 
-    const code = await runCommand("docker", [
-      "run",
-      "--rm",
-      "-v",
-      `${outDir}:/out`,
-      imageName,
+    const code = await runCommand(ytdlpBin, [
       "--no-progress",
       "--ignore-errors",
       "--continue",
@@ -170,7 +87,7 @@ async function main() {
       "0",
       "--add-metadata",
       "-o",
-      "/out/%(title)s.%(ext)s",
+      path.join(outDir, "%(title)s.%(ext)s"),
       url,
     ]);
